add (lab6): finish virtual memory and run successfully on vm.img

+++ src/cpio.c
@@ -1,8 +1,8 @@
 #include "cpio.h"
 
 const unsigned long HEADER_SIZE = sizeof(struct cpio_newc_header);
-uint32_t cpio_addr;
-uint32_t cpio_end;
+uint64_t cpio_addr;
+uint64_t cpio_end;
 
 char magic[6] = "070701";
 char header_magic[6];
